Move AnalyticsAdapter under libraries
