+++ libraries/analyticsAdapter/examples/example.js
@@ -2,7 +2,7 @@
  * example.js - analytics adapter for Example Analytics Library example
  */
 
-import adapter from '../../AnalyticsAdapter.js';
+import adapter from '../AnalyticsAdapter.js';
 
 export default adapter(
   {
