+++ libraries/analyticsAdapter/examples/example2.js
@@ -5,7 +5,7 @@ import { ajax } from '../../../src/ajax.js';
  * example2.js - analytics adapter for Example2 Analytics Endpoint example
  */
 
-import adapter from '../../AnalyticsAdapter.js';
+import adapter from '../AnalyticsAdapter.js';
 
 const url = 'https://httpbin.org/post';
 const analyticsType = 'endpoint';
