+++ libraries/analyticsAdapter/examples/libraries/example.js
