+++ libraries/analyticsAdapter/examples/libraries/example2.js
