+++ modules/adomikAnalyticsAdapter.js
@@ -1,4 +1,4 @@
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import CONSTANTS from '../src/constants.json';
 import adapterManager from '../src/adapterManager.js';
 import {logInfo} from '../src/utils.js';
