+++ modules/adxpremiumAnalyticsAdapter.js
@@ -1,6 +1,6 @@
 import {deepClone, logError, logInfo} from '../src/utils.js';
 import {ajax} from '../src/ajax.js';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import adapterManager from '../src/adapterManager.js';
 import CONSTANTS from '../src/constants.json';
 import {includes} from '../src/polyfill.js';
