+++ modules/atsAnalyticsAdapter.js
@@ -1,5 +1,5 @@
 import { logError, logInfo } from '../src/utils.js';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import CONSTANTS from '../src/constants.json';
 import adaptermanager from '../src/adapterManager.js';
 import {ajax} from '../src/ajax.js';
