+++ modules/invisiblyAnalyticsAdapter.js
@@ -2,7 +2,7 @@
  * invisiblyAdapterAdapter.js - analytics adapter for Invisibly
  */
 import { ajaxBuilder } from '../src/ajax.js';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import adapterManager from '../src/adapterManager.js';
 
 import { generateUUID, logInfo } from '../src/utils.js';
