+++ modules/konduitAnalyticsAdapter.js
@@ -1,6 +1,6 @@
 import { parseSizesInput, logError, uniques } from '../src/utils.js';
 import { ajax } from '../src/ajax.js';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import adapterManager from '../src/adapterManager.js';
 import { targeting } from '../src/targeting.js';
 import { config } from '../src/config.js';
