+++ modules/medianetAnalyticsAdapter.js
@@ -11,7 +11,7 @@ import {
   uniques,
   getHighestCpm
 } from '../src/utils.js';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import adapterManager from '../src/adapterManager.js';
 import CONSTANTS from '../src/constants.json';
 import {ajax} from '../src/ajax.js';
