+++ modules/ooloAnalyticsAdapter.js
@@ -1,6 +1,6 @@
 import { _each, deepClone, pick, deepSetValue, logError, logInfo } from '../src/utils.js';
 import { getOrigin } from '../libraries/getOrigin/index.js';
-import adapter from '../src/AnalyticsAdapter.js'
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js'
 import adapterManager from '../src/adapterManager.js'
 import CONSTANTS from '../src/constants.json'
 import { ajax } from '../src/ajax.js'
