+++ modules/openxAnalyticsAdapter.js
@@ -13,7 +13,7 @@ import {
   parseSizesInput,
   uniques
 } from '../src/utils.js';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import CONSTANTS from '../src/constants.json';
 import adapterManager from '../src/adapterManager.js';
 import {ajax} from '../src/ajax.js';
