+++ modules/optimonAnalyticsAdapter.js
@@ -8,7 +8,7 @@
 *
 */
 
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import adapterManager from '../src/adapterManager.js';
 
 const optimonAnalyticsAdapter = adapter({
