+++ modules/pubmaticAnalyticsAdapter.js
@@ -1,5 +1,5 @@
 import { _each, pick, logWarn, isStr, isArray, logError } from '../src/utils.js';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import adapterManager from '../src/adapterManager.js';
 import CONSTANTS from '../src/constants.json';
 import { ajax } from '../src/ajax.js';
