+++ modules/pubstackAnalyticsAdapter.js
@@ -1,4 +1,4 @@
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import adapterManager from '../src/adapterManager.js';
 
 const pubstackAnalytics = adapter({
