+++ modules/pulsepointAnalyticsAdapter.js
@@ -2,7 +2,7 @@
  * pulsepoint.js - Analytics Adapter for PulsePoint
  */
 
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import adapterManager from '../src/adapterManager.js';
 
 var pulsepointAdapter = adapter({
