+++ modules/scaleableAnalyticsAdapter.js
@@ -2,7 +2,7 @@
 
 import { ajax } from '../src/ajax.js';
 import CONSTANTS from '../src/constants.json';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import adapterManager from '../src/adapterManager.js';
 import { logMessage } from '../src/utils.js';
