+++ modules/sigmoidAnalyticsAdapter.js
@@ -1,7 +1,7 @@
 /* Sigmoid Analytics Adapter for prebid.js v1.1.0-pre
 Updated : 2018-03-28 */
 import {includes} from '../src/polyfill.js';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import CONSTANTS from '../src/constants.json';
 import adapterManager from '../src/adapterManager.js';
 import {getStorageManager} from '../src/storageManager.js';
