+++ modules/yieldoneAnalyticsAdapter.js
@@ -1,6 +1,6 @@
 import { isArray, deepClone } from '../src/utils.js';
 import {ajax} from '../src/ajax.js';
-import adapter from '../src/AnalyticsAdapter.js';
+import adapter from '../libraries/analyticsAdapter/AnalyticsAdapter.js';
 import CONSTANTS from '../src/constants.json';
 import adapterManager from '../src/adapterManager.js';
 import { targeting } from '../src/targeting.js';
